Renamed the Android and iOS apps to 'EZNet'
