+++ android/app/BUCK
@@ -46,13 +46,13 @@ android_library(
 
 android_build_config(
   name = 'build_config',
-  package = 'com.reactnativeproject',
+  package = 'org.gdgsrilanka.codelanka.eznet',
 )
 
 android_resource(
   name = 'res',
   res = 'src/main/res',
-  package = 'com.reactnativeproject',
+  package = 'org.gdgsrilanka.codelanka.eznet',
 )
 
 android_binary(
