+++ android/app/src/main/java/org/gdgsrilanka/codelanka/eznet/MainActivity.java
@@ -1,4 +1,4 @@
-package com.reactnativeproject;
+package org.gdgsrilanka.codelanka.eznet;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "ReactNativeProject";
+        return "EZNet";
     }
 }
