+++ android/app/src/main/java/org/gdgsrilanka/codelanka/eznet/MainApplication.java
@@ -1,4 +1,4 @@
-package com.reactnativeproject;
+package org.gdgsrilanka.codelanka.eznet;
 
 import android.app.Application;
 import android.util.Log;
@@ -10,6 +10,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
+
 
 import java.util.Arrays;
 import java.util.List;
