+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "ReactNativeProject",
-  "displayName": "ReactNativeProject"
+  "name": "EZNetApp",
+  "displayName": "EZNetApp"
 }
