+++ ios/EZNet/Images.xcassets/AppIcon.appiconset/Contents.json
