+++ ios/EZNetTests/EZNetTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface ReactNativeProjectTests : XCTestCase
+@interface EZNetTests : XCTestCase
 
 @end
 
-@implementation ReactNativeProjectTests
+@implementation EZNetTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
