+++ ios/EZNetTests/Info.plist
