+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ReactNativeProject",
+  "name": "EZNet",
   "version": "0.0.1",
   "private": true,
   "scripts": {
